modified:   .gitignore 	new file:   app/src/api/auth/login.ts 	renamed:    app/src/api/api.ts -> app/src/api/auth/postRequest.ts 	new file:   app/src/api/auth/signup.ts 	modified:   app/src/app/login/page.tsx 	modified:   app/src/app/signup/page.tsx 	modified:   app/src/components/forms/LoginForm.tsx 	modified:   app/src/components/forms/SignupForm.tsx
